It's almost dumping things correctly
The frame seems... suspect at best but I don't know...
The first frame of the movie I'm using is a whole black square so...
You could encode that with a dumb RLE encoding using a few bytes at best.

Anyway, it's generating the files correctly, now I just need to be sure
to dump the frames correctly.

And prepare a "dumped regs analyzer". This... will be hard.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>