h264_dump_displayer: Added the offset of each big member.
Turns out that offset_of cannot get the address of the bitfield.
That's... obvious in some sense, less in others.

But yeah, address relates to "bytes" and bits are only the content
of one "byte", but cannot be addressed as such...

Well.

I also started to dump more frames and, this time, the size of
each frame seemed to be more coherent as time goes in the video.
Basically the first 1080p black square takes nothing, but moving
animation have higher sizes.

Signed-off-by: Miouyouyou (Myy) <myy@miouyouyou.fr>